+++ keyboards/leafcutterlabs/bigknobv2/bigknobv2.c
@@ -1,4 +1,4 @@
-#include "bigKNOBv2.h"
+#include "bigknobv2.h"
 
 void matrix_init_kb(void) {
 	matrix_init_user();
